refactor(java): move isLatin from StringSerializer to StringUtils
