+++ java/fury-core/src/main/java/org/apache/fury/util/StringUtils.java
@@ -22,11 +22,25 @@ package org.apache.fury.util;
 import java.util.HashMap;
 import java.util.Map;
 import java.util.Random;
+import org.apache.fury.memory.Platform;
 
 public class StringUtils {
   private static final char[] BASE16_CHARS2 = {
     '0', '1', '2', '3', '4', '5', '6', '7', '8', '9', 'a', 'b', 'c', 'd', 'e', 'f'
   };
+  private static final long MULTI_CHARS_NON_LATIN_MASK;
+
+  static {
+    if (Platform.IS_LITTLE_ENDIAN) {
+      // latin chars will be 0xXX,0x00;0xXX,0x00 in byte order;
+      // Using 0x00,0xff(0xff00) to clear latin bits.
+      MULTI_CHARS_NON_LATIN_MASK = 0xff00ff00ff00ff00L;
+    } else {
+      // latin chars will be 0x00,0xXX;0x00,0xXX in byte order;
+      // Using 0x00,0xff(0x00ff) to clear latin bits.
+      MULTI_CHARS_NON_LATIN_MASK = 0x00ff00ff00ff00ffL;
+    }
+  }
 
   /** Converts a bytes array into a hexadecimal string. */
   public static String encodeHexString(final byte[] data) {
@@ -248,5 +262,31 @@ public class StringUtils {
     }
 
     return builder.toString();
+  }
+
+  public static boolean isLatin(char[] chars) {
+    int numChars = chars.length;
+    int vectorizedLen = numChars >> 2;
+    int vectorizedChars = vectorizedLen << 2;
+    int endOffset = Platform.CHAR_ARRAY_OFFSET + (vectorizedChars << 1);
+    boolean isLatin = true;
+    for (int offset = Platform.CHAR_ARRAY_OFFSET; offset < endOffset; offset += 8) {
+      // check 4 chars in a vectorized way, 4 times faster than scalar check loop.
+      // See benchmark in CompressStringSuite.latinSuperWordCheck.
+      long multiChars = Platform.getLong(chars, offset);
+      if ((multiChars & MULTI_CHARS_NON_LATIN_MASK) != 0) {
+        isLatin = false;
+        break;
+      }
+    }
+    if (isLatin) {
+      for (int i = vectorizedChars; i < numChars; i++) {
+        if (chars[i] > 0xFF) {
+          isLatin = false;
+          break;
+        }
+      }
+    }
+    return isLatin;
   }
 }
