+++ java/fury-core/src/test/java/org/apache/fury/serializer/StringSerializerTest.java
@@ -21,8 +21,6 @@ package org.apache.fury.serializer;
 
 import static org.apache.fury.serializer.StringSerializer.newBytesStringZeroCopy;
 import static org.testng.Assert.assertEquals;
-import static org.testng.Assert.assertFalse;
-import static org.testng.Assert.assertTrue;
 
 import java.lang.reflect.Field;
 import java.nio.ByteBuffer;
@@ -288,87 +286,6 @@ public class StringSerializerTest extends FuryTestBase {
       assertEquals(stringSerializer.readJavaString(buffer), latinStr);
       assertEquals(buffer.writerIndex(), buffer.readerIndex());
     }
-  }
-
-  @Test(dataProvider = "endian")
-  public void testVectorizedLatinCheckAlgorithm(boolean endian) {
-    // assertTrue(isLatin("Fury".toCharArray(), endian));
-    // assertTrue(isLatin(StringUtils.random(8 * 10).toCharArray(), endian));
-    // test unaligned
-    assertTrue(isLatin((StringUtils.random(8 * 10) + "1").toCharArray(), endian));
-    assertTrue(isLatin((StringUtils.random(8 * 10) + "12").toCharArray(), endian));
-    assertTrue(isLatin((StringUtils.random(8 * 10) + "123").toCharArray(), endian));
-    assertFalse(isLatin("你好, Fury".toCharArray(), endian));
-    assertFalse(isLatin((StringUtils.random(8 * 10) + "你好").toCharArray(), endian));
-    assertFalse(isLatin((StringUtils.random(8 * 10) + "1你好").toCharArray(), endian));
-  }
-
-  private boolean isLatin(char[] chars, boolean isLittle) {
-    boolean reverseBytes =
-        (Platform.IS_LITTLE_ENDIAN && !isLittle) || (!Platform.IS_LITTLE_ENDIAN && !isLittle);
-    if (reverseBytes) {
-      for (int i = 0; i < chars.length; i++) {
-        chars[i] = Character.reverseBytes(chars[i]);
-      }
-    }
-    long mask;
-    if (isLittle) {
-      // latin chars will be 0xXX,0x00;0xXX,0x00 in byte order;
-      // Using 0x00,0xff(0xff00) to clear latin bits.
-      mask = 0xff00ff00ff00ff00L;
-    } else {
-      // latin chars will be 0x00,0xXX;0x00,0xXX in byte order;
-      // Using 0x00,0xff(0x00ff) to clear latin bits.
-      mask = 0x00ff00ff00ff00ffL;
-    }
-    int numChars = chars.length;
-    int vectorizedLen = numChars >> 2;
-    int vectorizedChars = vectorizedLen << 2;
-    int endOffset = Platform.CHAR_ARRAY_OFFSET + (vectorizedChars << 1);
-    boolean isLatin = true;
-    for (int offset = Platform.CHAR_ARRAY_OFFSET; offset < endOffset; offset += 8) {
-      // check 4 chars in a vectorized way, 4 times faster than scalar check loop.
-      long multiChars = Platform.getLong(chars, offset);
-      if ((multiChars & mask) != 0) {
-        isLatin = false;
-        break;
-      }
-    }
-    if (isLatin) {
-      for (int i = vectorizedChars; i < numChars; i++) {
-        char c = chars[i];
-        if (reverseBytes) {
-          c = Character.reverseBytes(c);
-        }
-        if (c > 0xFF) {
-          isLatin = false;
-          break;
-        }
-      }
-    }
-    return isLatin;
-  }
-
-  @Test
-  public void testLatinCheck() {
-    assertTrue(StringSerializer.isLatin("Fury".toCharArray()));
-    assertTrue(StringSerializer.isLatin(StringUtils.random(8 * 10).toCharArray()));
-    // test unaligned
-    assertTrue(StringSerializer.isLatin((StringUtils.random(8 * 10) + "1").toCharArray()));
-    assertTrue(StringSerializer.isLatin((StringUtils.random(8 * 10) + "12").toCharArray()));
-    assertTrue(StringSerializer.isLatin((StringUtils.random(8 * 10) + "123").toCharArray()));
-    assertFalse(StringSerializer.isLatin("你好, Fury".toCharArray()));
-    assertFalse(StringSerializer.isLatin((StringUtils.random(8 * 10) + "你好").toCharArray()));
-    assertFalse(StringSerializer.isLatin((StringUtils.random(8 * 10) + "1你好").toCharArray()));
-    assertFalse(StringSerializer.isLatin((StringUtils.random(11) + "你").toCharArray()));
-    assertFalse(StringSerializer.isLatin((StringUtils.random(10) + "你好").toCharArray()));
-    assertFalse(StringSerializer.isLatin((StringUtils.random(9) + "性能好").toCharArray()));
-    assertFalse(StringSerializer.isLatin("\u1234".toCharArray()));
-    assertFalse(StringSerializer.isLatin("a\u1234".toCharArray()));
-    assertFalse(StringSerializer.isLatin("ab\u1234".toCharArray()));
-    assertFalse(StringSerializer.isLatin("abc\u1234".toCharArray()));
-    assertFalse(StringSerializer.isLatin("abcd\u1234".toCharArray()));
-    assertFalse(StringSerializer.isLatin("Javaone Keynote\u1234".toCharArray()));
   }
 
   @Test
